pyproject.toml
[project]
name = "zrc-bot"
version = "0.1.0"
description = "zrc-bot"
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "aiocache==0.12.3",
    "aiofiles==24.1.0",
    "aiosqlite==0.21.0",
    "alembic==1.16.4",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "arclet-alconna==1.8.40",
    "arclet-alconna-tools==0.7.11",
    "asgiref==3.9.1",
    "async-asgi-testclient==1.4.11",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "cloudscraper==1.2.71",
    "colorama==0.4.6",
    "cryptography==45.0.6",
    "dowhen==0.1.0",
    "exceptiongroup==1.3.0",
    "fastapi==0.116.1",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "importlib-metadata==8.7.0",
    "iniconfig==2.1.0",
    "jaconv==0.4.0",
    "jinja2==3.1.6",
    "loguru==0.7.3",
    "lxml==5.4.0",
    "maimai-ffi==0.6.0",
    "maimai-py==1.3.5",
    "mako==1.3.10",
    "markdown==3.8.2",
    "markupsafe==3.0.2",
    "msgpack==1.1.1",
    "multidict==6.6.4",
    "nepattern==0.7.7",
    "nonebot-adapter-onebot==2.4.6",
    "nonebot-plugin-alconna==0.59.4",
    "nonebot-plugin-htmlrender==0.6.6",
    "nonebot-plugin-llm-helper>=0.0.4",
    "nonebot-plugin-localstore==0.7.4",
    "nonebot-plugin-orm==0.8.2",
    "nonebot-plugin-waiter==0.8.1",
    "nonebot2==2.4.3",
    "nonebug==0.4.3",
    "nonestorage==0.1.0",
    "packaging==25.0",
    "pillow==11.3.0",
    "playwright==1.54.0",
    "pluggy==1.6.0",
    "propcache==0.3.2",
    "pycparser==2.22",
    "pycryptodome==3.23.0",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pyee==13.0.0",
    "pygments==2.19.2",
    "pygtrie==2.5.0",
    "pymdown-extensions==10.16.1",
    "pyparsing==3.2.3",
    "pytest==8.4.1",
    "pytest-asyncio==1.1.0",
    "python-dotenv==1.1.1",
    "python-markdown-math==0.9",
    "pyyaml==6.0.2",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.43",
    "starlette==0.47.2",
    "tarina==0.6.8",
    "tenacity==9.1.2",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "wahlap-mai-ass-expander",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
    "win32-setctime==1.2.0",
    "yarl==1.20.1",
    "zipp==3.23.0",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]
plugins = [
    "nonebot_plugin_orm",
    "nonebot_plugin_localstore",
    "nonebot_plugin_alconna",
    "nonebot_plugin_htmlrender",
]
plugin_dirs = ["src/plugins"]
builtin_plugins = []

[tool.ruff]
ignore = ["E402"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[tool.uv.sources]
wahlap-mai-ass-expander = { git = "ssh://git@github.com/Nice-Remi-Cluster/Wahlap-Mai-Ass-Expander.git", rev = "494b97fcc33dcb58eb7772664ab811ed05017bed" }
